Trait pairing::ff::PrimeFieldRepr

pub trait PrimeFieldRepr: 'static + Send + Copy + Sync + Eq + Ord + Clone + Default + AsRef<[u64]> + AsMut<[u64]> + From<u64> + Debug + Display + Rand {
    fn sub_noborrow(&mut self, other: &Self);

    fn add_nocarry(&mut self, other: &Self);

    fn num_bits(&self) -> u32;

    fn is_zero(&self) -> bool;

    fn is_odd(&self) -> bool;

    fn is_even(&self) -> bool;

    fn div2(&mut self);

    fn shr(&mut self, amt: u32);

    fn mul2(&mut self);

    fn shl(&mut self, amt: u32);

    fn write_be<W>(&self, writer: W) -> Result<(), Error>
    where
        W: Write,
    { ... }

    fn read_be<R>(&mut self, reader: R) -> Result<(), Error>
    where
        R: Read,
    { ... }

    fn write_le<W>(&self, writer: W) -> Result<(), Error>
    where
        W: Write,
    { ... }

    fn read_le<R>(&mut self, reader: R) -> Result<(), Error>
    where
        R: Read,
    { ... }
}

This trait represents a wrapper around a biginteger which can encode any element of a particular prime field. It is a smart wrapper around a sequence of u64 limbs, least-significant digit first.

Required methods

fn sub_noborrow(&mut self, other: &Self)

Subtract another represetation from this one.

fn add_nocarry(&mut self, other: &Self)

Add another representation to this one.

fn num_bits(&self) -> u32

Compute the number of bits needed to encode this number. Always a multiple of 64.

fn is_zero(&self) -> bool

Returns true iff this number is zero.

fn is_odd(&self) -> bool

Returns true iff this number is odd.

fn is_even(&self) -> bool

Returns true iff this number is even.

fn div2(&mut self)

Performs a rightwise bitshift of this number, effectively dividing it by 2.

fn shr(&mut self, amt: u32)

Performs a rightwise bitshift of this number by some amount.

fn mul2(&mut self)

Performs a leftwise bitshift of this number, effectively multiplying it by 2. Overflow is ignored.

fn shl(&mut self, amt: u32)

Performs a leftwise bitshift of this number by some amount.

Loading content...
Provided methods

fn write_be<W>(&self, writer: W) -> Result<(), Error> where
W: Write,

Writes this PrimeFieldRepr as a big endian integer.

fn read_be<R>(&mut self, reader: R) -> Result<(), Error> where
R: Read,

Reads a big endian integer into this representation.

fn write_le<W>(&self, writer: W) -> Result<(), Error> where
W: Write,

Writes this PrimeFieldRepr as a little endian integer.

fn read_le<R>(&mut self, reader: R) -> Result<(), Error> where
R: Read,

Reads a little endian integer into this representation.

Loading content...