Trait pairing::EncodedPoint

pub trait EncodedPoint: Sized + Send + Sync + AsRef<[u8]> + AsMut<[u8]> + Clone + Copy + 'static {
    type Affine: CurveAffine;
    fn empty() -> Self;

    fn size() -> usize;

    fn into_affine(&self) -> Result<Self::Affine, GroupDecodingError>;

    fn into_affine_unchecked(&self) -> Result<Self::Affine, GroupDecodingError>;

    fn from_affine(affine: Self::Affine) -> Self;
}

An encoded elliptic curve point, which should essentially wrap a [u8; N].

Associated Types

type Affine: CurveAffine

Loading content...
Required methods

fn empty() -> Self

Creates an empty representation.

fn size() -> usize

Returns the number of bytes consumed by this representation.

fn into_affine(&self) -> Result<Self::Affine, GroupDecodingError>

Converts an EncodedPoint into a CurveAffine element, if the encoding represents a valid element.

fn into_affine_unchecked(&self) -> Result<Self::Affine, GroupDecodingError>

Converts an EncodedPoint into a CurveAffine element, without guaranteeing that the encoding represents a valid element. This is useful when the caller knows the encoding is valid already.

If the encoding is invalid, this can break API invariants, so caution is strongly encouraged.

fn from_affine(affine: Self::Affine) -> Self

Creates an EncodedPoint from an affine point, as long as the point is not the point at infinity.

Loading content...