Crate pairing

A library for working with pairing-friendly curves.

Re-exports

pub use group;

Traits

Engine

An “engine” is a collection of types (fields, elliptic curve groups, etc.) with well-defined relationships. In particular, the G1/G2 curve groups are of prime order r, and are equipped with a bilinear pairing function.

MillerLoopResult

Represents results of a Miller loop, one of the most expensive portions of the pairing function.

MultiMillerLoop

An engine that can compute sums of pairings in an efficient way.

PairingCurveAffine

Affine representation of an elliptic curve point that can be used to perform pairings.