Crate pairing

Modules

bls12_381

bn256

ff

Structs

Wnaf
A "w-ary non-adjacent form" exponentiation context.

Enums

GroupDecodingError
An error that may occur when trying to decode an EncodedPoint.

Traits

CurveAffine
Affine representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

CurveProjective
Projective representation of an elliptic curve point guaranteed to be in the correct prime order subgroup.

EncodedPoint
An encoded elliptic curve point, which should essentially wrap a [u8; N].

Engine
An "engine" is a collection of types (fields, elliptic curve groups, etc.) with well-defined relationships. In particular, the G1/G2 curve groups are of prime order r, and are equipped with a bilinear pairing function.