Crate pairing

A library for working with pairing-friendly curves.

Modules

bls12_381

An implementation of the BLS12-381 pairing-friendly elliptic curve construction.

Traits

Engine

An “engine” is a collection of types (fields, elliptic curve groups, etc.) with well-defined relationships. In particular, the G1/G2 curve groups are of prime order r, and are equipped with a bilinear pairing function.

PairingCurveAffine

Affine representation of an elliptic curve point that can be used to perform pairings.